Moved the Enums to enum folder for classes analyzers
