+++ src/backend/KnowledgeNetwork.Domains.Code/Analyzers/Classes/Utilities/ClassNodeFactory.cs
@@ -1,5 +1,6 @@
 using KnowledgeNetwork.Domains.Code.Analyzers.Classes.Abstractions;
 using KnowledgeNetwork.Domains.Code.Models.Classes;
+using KnowledgeNetwork.Domains.Code.Models.Classes.ClassEnums;
 using KnowledgeNetwork.Domains.Code.Models.Common;
 using Microsoft.CodeAnalysis;
 using Microsoft.CodeAnalysis.CSharp;
