+++ src/backend/KnowledgeNetwork.Domains.Code/Models/Classes/ClassEnums/ClassType.cs
@@ -1,0 +1,15 @@
+﻿namespace KnowledgeNetwork.Domains.Code.Models.Classes.ClassEnums;
+
+/// <summary>
+/// Types of classes/types
+/// </summary>
+public enum ClassType
+{
+    Class,
+    Interface,
+    Struct,
+    Enum,
+    Record,
+    RecordStruct,
+    Delegate
+}
