+++ src/backend/KnowledgeNetwork.Domains.Code/Models/Classes/ClassEnums/ClassVisibility.cs
@@ -1,0 +1,14 @@
+﻿namespace KnowledgeNetwork.Domains.Code.Models.Classes.ClassEnums;
+
+/// <summary>
+/// Visibility levels for classes
+/// </summary>
+public enum ClassVisibility
+{
+    Private,
+    Protected,
+    Internal,
+    Public,
+    ProtectedInternal,
+    PrivateProtected
+}
