+++ src/backend/KnowledgeNetwork.Domains.Code/Models/Classes/InheritanceEdge.cs
@@ -1,3 +1,4 @@
+using KnowledgeNetwork.Domains.Code.Models.Classes.ClassEnums;
 using KnowledgeNetwork.Domains.Code.Models.Common;
 
 namespace KnowledgeNetwork.Domains.Code.Models.Classes;
@@ -56,30 +57,4 @@ public class InheritanceEdge
     /// Additional metadata
     /// </summary>
     public Dictionary<string, object> Metadata { get; set; } = new();
-}
-
-/// <summary>
-/// Types of inheritance relationships
-/// </summary>
-public enum InheritanceType
-{
-    /// <summary>
-    /// Class to class inheritance
-    /// </summary>
-    Class,
-
-    /// <summary>
-    /// Class to abstract class inheritance
-    /// </summary>
-    AbstractClass,
-
-    /// <summary>
-    /// Record to record inheritance
-    /// </summary>
-    Record,
-
-    /// <summary>
-    /// Interface to interface inheritance
-    /// </summary>
-    Interface
 }
