+++ src/backend/KnowledgeNetwork.Domains.Code/Models/Classes/InterfaceImplementationEdge.cs
@@ -1,3 +1,4 @@
+using KnowledgeNetwork.Domains.Code.Models.Classes.ClassEnums;
 using KnowledgeNetwork.Domains.Code.Models.Common;
 
 namespace KnowledgeNetwork.Domains.Code.Models.Classes;
@@ -71,61 +72,4 @@ public class InterfaceImplementationEdge
     /// Additional metadata
     /// </summary>
     public Dictionary<string, object> Metadata { get; set; } = new();
-}
-
-/// <summary>
-/// Types of interface implementation
-/// </summary>
-public enum InterfaceImplementationType
-{
-    /// <summary>
-    /// Direct implementation by the class
-    /// </summary>
-    Direct,
-
-    /// <summary>
-    /// Inherited from base class
-    /// </summary>
-    Inherited,
-
-    /// <summary>
-    /// Both direct and inherited
-    /// </summary>
-    DirectAndInherited
-}
-
-/// <summary>
-/// Information about an interface method implementation
-/// </summary>
-public class InterfaceMethodImplementation
-{
-    /// <summary>
-    /// Name of the interface method
-    /// </summary>
-    public string MethodName { get; set; } = string.Empty;
-
-    /// <summary>
-    /// Signature of the interface method
-    /// </summary>
-    public string MethodSignature { get; set; } = string.Empty;
-
-    /// <summary>
-    /// Name of the implementing method in the class
-    /// </summary>
-    public string ImplementingMethodName { get; set; } = string.Empty;
-
-    /// <summary>
-    /// Whether this is an explicit implementation
-    /// </summary>
-    public bool IsExplicit { get; set; }
-
-    /// <summary>
-    /// Whether the implementation is in the current class or inherited
-    /// </summary>
-    public bool IsInherited { get; set; }
-
-    /// <summary>
-    /// Source location of the implementation
-    /// </summary>
-    public CSharpLocationInfo? ImplementationLocation { get; set; }
 }
